refactor: extract Windows hamburger menu to win_hamburger.rs and fix cfg guards
