+++ desktop/src/components/app.rs
@@ -7,7 +7,9 @@ mod shortcut_overlay;
 
 use dioxus::desktop::tao::dpi::{LogicalPosition, LogicalSize};
 use dioxus::desktop::tao::event::{DeviceEvent, ElementState, Event as TaoEvent, WindowEvent};
-use dioxus::desktop::{use_muda_event_handler, use_wry_event_handler, window};
+#[cfg(target_os = "macos")]
+use dioxus::desktop::use_muda_event_handler;
+use dioxus::desktop::{use_wry_event_handler, window};
 use dioxus::document;
 use dioxus::prelude::*;
 use dioxus_core::use_drop;
@@ -25,6 +27,7 @@ use super::tab::TabBar;
 use crate::assets::MAIN_SCRIPT;
 use crate::drag;
 use crate::events::{ActiveDragUpdate, ACTIVE_DRAG_UPDATE};
+#[cfg(target_os = "macos")]
 use crate::menu;
 use crate::state::{AppState, PersistedState, Tab};
 use crate::theme::Theme;
@@ -142,6 +145,7 @@ pub fn App(
     setup_keybinding_engine(state, shortcut_overlay_visibility);
 
     // Handle menu events (only state-dependent events, not global ones)
+    #[cfg(target_os = "macos")]
     use_muda_event_handler(move |event| {
         // Only handle state-dependent events
         menu::handle_menu_event_with_state(event, &mut state);
