+++ desktop/src/menu.rs
@@ -1,5 +1,3 @@
-#![allow(dead_code)]
-
 use dioxus::document;
 use dioxus::prelude::{spawn, ReadableExt, WritableExt};
 use dioxus_desktop::muda::accelerator::Accelerator;
