+++ desktop/src/window.rs
@@ -51,10 +51,12 @@ pub use child::{
     close_child_windows_for_last_focused, close_child_windows_for_parent,
     open_or_focus_image_window, open_or_focus_math_window, open_or_focus_mermaid_window,
 };
+#[cfg(target_os = "macos")]
+pub use main::has_any_main_windows;
 pub use main::{
     close_all_main_windows, create_main_window_config, create_main_window_sync,
-    create_main_window_sync_with_tabs, get_any_main_window, has_any_main_windows,
-    register_main_window, register_window_state, shutdown_all_windows, unregister_window_state,
+    create_main_window_sync_with_tabs, get_any_main_window, register_main_window,
+    register_window_state, shutdown_all_windows, unregister_window_state,
     update_last_focused_window, CreateMainWindowConfigParams,
 };
 pub use preview::{
